move chatbot-api to backend/chat-core
